requirements.txt
asgiref==3.2.10
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
Django==3.1.2
djangorestframework==3.12.1
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
nltk==3.5
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.11
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.51.0
